WIP Enabling bluetooth
Adds a patch for missing kernel configuration along with a list of
kernel modules that are missing by default.

Still not working as of yet, though another user has confirmed that
they're able to get bluetooth working by copying the entire Arch kernel
config, so we must just be missing some lines of kernel config.

Tries adding the INTEL and SERDEV bluetooth config params, but to no
avail yet.